+++ packages/expo-server/src/rendering.ts
@@ -7,8 +7,11 @@ import type { AssetInfo, GetStaticContentOptions } from './manifest';
  * {@link import('@expo/router-server/src/static/renderStaticContent')}
  */
 export interface ServerRenderModule {
-  /** {@link import('@expo/router-server/src/static/renderStaticContent').getStaticContent} */
-  getStaticContent(location: URL, options?: GetStaticContentOptions): Promise<string>;
+  /** {@link import('@expo/router-server/src/static/renderStaticContent').getStreamingContent} */
+  getStreamingContent(
+    location: URL,
+    options?: GetStaticContentOptions
+  ): Promise<ReadableStream<Uint8Array>>;
 }
 
 export interface RenderOptions {
@@ -16,7 +19,10 @@ export interface RenderOptions {
   assets?: AssetInfo;
 }
 
-export type SsrRenderFn = (request: Request, options?: RenderOptions) => Promise<string>;
+export type SsrRenderFn = (
+  request: Request,
+  options?: RenderOptions
+) => Promise<ReadableStream<Uint8Array>>;
 
 /** Module exported from loader bundle, typically `_expo/loaders/[ROUTE].js` */
 export interface LoaderModule {